+++ Frends.As4.ValidateAndParsePayload/Frends.As4.ValidateAndParsePayload.Tests/ErrorHandlerTest.cs
@@ -1,10 +1,10 @@
 using System;
 using System.Threading;
+using System.Threading.Tasks;
 using NUnit.Framework;
 
 namespace Frends.As4.ValidateAndParsePayload.Tests;
 
-// TODO: Adjust the test to use a real invalid Input scenario (e.g., missing or malformed data)
 [TestFixture]
 internal class ErrorHandlerTest : TestBase
 {
@@ -13,17 +13,26 @@ internal class ErrorHandlerTest : TestBase
     [Test]
     public void Should_Throw_Error_When_ThrowErrorOnFailure_Is_True()
     {
-        var ex = Assert.Throws<Exception>((Action)(() =>
-           As4.ValidateAndParsePayload(DefaultInput(), DefaultConnection(), DefaultOptions(), CancellationToken.None)));
+        var options = DefaultOptions();
+        options.ThrowErrorOnFailure = true;
+
+        var ex = Assert.ThrowsAsync<Exception>((Func<Task>)(async () =>
+            await As4.ValidateAndParsePayload(
+                EmptyInput(),
+                DefaultConnection(),
+                options,
+                CancellationToken.None)));
+
         Assert.That(ex, Is.Not.Null);
     }
 
     [Test]
-    public void Should_Return_Failed_Result_When_ThrowErrorOnFailure_Is_False()
+    public async Task Should_Return_Failed_Result_When_ThrowErrorOnFailure_Is_False()
     {
         var options = DefaultOptions();
         options.ThrowErrorOnFailure = false;
-        var result = As4.ValidateAndParsePayload(DefaultInput(), DefaultConnection(), options, CancellationToken.None);
+        var result =
+            await As4.ValidateAndParsePayload(EmptyInput(), DefaultConnection(), options, CancellationToken.None);
         Assert.That(result.Success, Is.False);
     }
 
@@ -31,9 +40,12 @@ internal class ErrorHandlerTest : TestBase
     public void Should_Use_Custom_ErrorMessageOnFailure()
     {
         var options = DefaultOptions();
+        options.ThrowErrorOnFailure = true;
         options.ErrorMessageOnFailure = CustomErrorMessage;
-        var ex = Assert.Throws<Exception>((Action)(() =>
-            As4.ValidateAndParsePayload(DefaultInput(), DefaultConnection(), options, CancellationToken.None)));
+
+        var ex = Assert.ThrowsAsync<Exception>((Func<Task>)(async () =>
+            await As4.ValidateAndParsePayload(EmptyInput(), DefaultConnection(), options, CancellationToken.None)));
+
         Assert.That(ex, Is.Not.Null);
         Assert.That(ex.Message, Contains.Substring(CustomErrorMessage));
     }