+++ Frends.As4.ValidateAndParsePayload/Frends.As4.ValidateAndParsePayload/Definitions/Connection.cs
@@ -1,19 +1,50 @@
 using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
+using Frends.As4.ValidateAndParsePayload.Attributes;
 
 namespace Frends.As4.ValidateAndParsePayload.Definitions;
 
 /// <summary>
 /// Connection parameters.
 /// </summary>
-// TODO: Remove this class if the task does not make connections
 public class Connection
 {
     /// <summary>
-    /// Connection string to the target service (e.g., database, API endpoint).
+    /// Forces signature validation when enabled.
     /// </summary>
-    /// <example>Host=127.0.0.1;Port=5432</example>
+    /// <example>true</example>
+    [DefaultValue(false)]
+    public bool RequireSigned { get; set; }
+
+    /// <summary>
+    /// Path to partner's public certificate (.cer/.pem) file used to validate the signature.
+    /// </summary>
+    /// <example>C:\Certs\partner.cer</example>
     [DisplayFormat(DataFormatString = "Text")]
-    [DefaultValue("")]
-    public string ConnectionString { get; set; } = string.Empty;
+    [RequiredIf(nameof(RequireSigned), true)]
+    public string PartnerCertificatePath { get; set; }
+
+    /// <summary>
+    /// Forces decryption when enabled.
+    /// </summary>
+    /// <example>true</example>
+    [DefaultValue(false)]
+    public bool RequireEncrypted { get; set; }
+
+    /// <summary>
+    /// Path to your own certificate file in .pfx format used for decryption/signature validation.
+    /// </summary>
+    /// <example>C:\Certs\mycompany.pfx</example>
+    [DisplayFormat(DataFormatString = "Text")]
+    [RequiredIfAny(true, nameof(RequireSigned), nameof(RequireEncrypted))]
+    public string OwnCertificatePath { get; set; }
+
+    /// <summary>
+    /// Password for your private key certificate.
+    /// </summary>
+    /// <example>mySecurePassword123</example>
+    [PasswordPropertyText]
+    [DisplayFormat(DataFormatString = "Text")]
+    [RequiredIfAny(true, nameof(RequireSigned), nameof(RequireEncrypted))]
+    public string OwnCertificatePassword { get; set; }
 }